perf: add power state tracking
Sometimes there might be such cases when system gives information
multiple times, or for multiple devices (like double battery for
laptop). In such cases we definitely don't want to process the same
power state event multiple times. So instead we will just track the
power state changes, and if the state is the same as the last one, we
will just ignore it.